backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
sqlalchemy[asyncio]>=2.0.25,<3.0.0
asyncpg>=0.29.0,<1.0.0
alembic>=1.13.0,<2.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
cryptography>=42.0.0,<44.0.0
httpx>=0.26.0,<1.0.0
websockets>=12.0,<14.0
python-multipart>=0.0.6,<1.0.0
redis>=5.0.0,<6.0.0
celery>=5.3.0,<6.0.0
anthropic>=0.40.0,<1.0.0
openai>=1.10.0,<2.0.0
ollama>=0.4.0,<1.0.0
structlog>=24.1.0,<25.0.0
python-dotenv>=1.0.0,<2.0.0
